imitator/imitator_yt_azure/requirements.txt
azure-functions
yt-dlp